--- src/test/java/me/zerosquare/simplemodel/Company.java
@@ -1,12 +1,0 @@
-package me.zerosquare.simplemodel;
-
-@Table(name = "companies")
-public class Company extends Model {
-  @Column
-  public Long id;
-
-  @Column
-  public String name;
-
-}
-
