+++ src/test/java/me/zerosquare/simplemodel/test/example/SimpleModelExample.java
@@ -1,13 +1,13 @@
-package me.zerosquare.simplemodel;
+package me.zerosquare.simplemodel.test.example;
 
 import java.sql.SQLException;
 import java.util.*;
-import java.lang.reflect.*;
 
 import me.zerosquare.simplemodel.internal.Connector;
 import me.zerosquare.simplemodel.internal.Logger;
+import me.zerosquare.simplemodel.test.Company;
+import me.zerosquare.simplemodel.test.Employee;
 import org.junit.*;
-import static org.junit.Assert.*;
 
 import me.zerosquare.simplemodel.*;
